Updated Encoder typekit

+++ components/lowlevel/encoder/types/EncoderDataType.hpp
@@ -5,13 +5,13 @@
 typedef struct
 {
 	/// Angle of the carousel [rad]
-	double theta;
+	float theta;
 	/// Sine and cosine of the angle
-	double sin_theta, cos_theta;
+	float sin_theta, cos_theta;
 	/// Raw speed [rad/s]
-	double omega;
+	float omega;
 	/// Filtered speed [rpm]
-	double omega_filt_rpm;
+	float omega_filt_rpm;
 	/// Time-stamps: trigger [ticks] and elapsed [sec]
 	double ts_trigger, ts_elapsed;
 } EncoderDataType;
